已完成：用户签到功能 回复功能 未完成：帖子功能 （数据库问题还在更改）

+++ src/main/java/com/eric/service/CommentService.java
@@ -16,7 +16,7 @@ public interface CommentService {
      * @param commentId
      * @return
      */
-    public int deleteComment(int commentId);
+    public int deleteComment(Long commentId);
 
     /**
      * 普通用户更新帖子
